uncorked-food/requirements.txt
dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.7.3.1
whitenoise==2.0.6
python-levenshtein==0.12.0
pandas==0.20.3
enum34>=1.1.6 ; python_version <= '2.7'
nltk==3.2.5